Forgetting a strong turn
Forgetting a strong turn

+++ tensorflow/examples/android/src/org/tensorflow/demo/TensorFlowYoloDetector.java
@@ -143,9 +143,9 @@ public class TensorFlowYoloDetector implements Classifier {
     bitmap.getPixels(intValues, 0, bitmap.getWidth(), 0, 0, bitmap.getWidth(), bitmap.getHeight());
 
     for (int i = 0; i < intValues.length; ++i) {
-      floatValues[i * 3 + 0] = (intValues[i] >> 16) / 255.0f;
-      floatValues[i * 3 + 1] = (intValues[i] >> 8) / 255.0f;
-      floatValues[i * 3 + 2] = intValues[i] / 255.0f;
+      floatValues[i * 3 + 0] = ((byte)(intValues[i] >> 16)) / 255.0f;
+      floatValues[i * 3 + 1] = ((byte)(intValues[i] >> 8)) / 255.0f;
+      floatValues[i * 3 + 2] = ((byte)intValues[i]) / 255.0f;
     }
     Trace.endSection(); // preprocessBitmap
